requirements.txt
absl-py==0.11.0
addict==2.3.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blender==1.4
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.4.0
gdown==3.12.2
grpcio==1.33.2
h5py==3.0.0
idna==2.10
imageio==2.9.0
imath==0.0.1
importlib-metadata==2.0.0
ipdb==0.13.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.0.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
msgpack==1.0.0
msgpack-numpy==0.4.7.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
notebook==6.1.5
numpy==1.19.4
open3d==0.11.1
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plyfile==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pyarrow==2.0.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyquaternion==0.9.9
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tabulate==0.8.7
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorpack==0.8.9
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.51.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
